+++ Server/character_creation.c
@@ -51,12 +51,12 @@ void check_new_character(int connection, const unsigned char *packet){
     char char_name[80]="";
     char password[80]="";
 
-    int packet_length=get_packet_length(packet);
+    size_t packet_length=get_packet_length(packet);
 
     //we know that there are always 3 bytes before the char name
     //and password and 7 bytes after, so we can extract this adding
     //3 to the packet start and subtracting 7 from packet end
-    strncpy(char_name_and_password, (char*)packet+3, (size_t)packet_length-7);
+    strncpy(char_name_and_password, (char*)packet+3, packet_length-7);
 
     //we know that the char name and password are separated by an
     //ascii space, so we can extract each separately by scanning
@@ -109,7 +109,7 @@ void add_new_character(int connection, const unsigned char *packet){
     }_packet_2;
 
     //calculate the total packet length
-    int packet_length=get_packet_length(packet);
+    size_t packet_length=get_packet_length(packet);
 
     //clear the structs
     memset(&_packet_1, '0', sizeof(_packet_1));
