+++ Server/idle_buffer2.cpp
@@ -63,7 +63,7 @@ typedef std::deque<data_> buffer_list_type;
 
 buffer_list_type idle_buffer2;
 
-void push_idle_buffer2(int connection, int process_type, const unsigned char *packet, int packet_len){
+void push_idle_buffer2(int connection, int process_type, const unsigned char *packet, size_t packet_len){
 
     /** public function - see header **/
